Remove _frozen_trial and _trial_state
